+++ src/main/java/com/example/demo/entity/User.java
@@ -32,7 +32,8 @@ public class User {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Long id;
+	private int id;
+	// private Long id;
 	
 	@Column(nullable=false)
 	private String name;
